+++ methods/docker/server.rb
@@ -1,14 +1,16 @@
+# frozen_string_literal: true
+
 # Docker server code
 
 # Configure Docker server
 
 def configure_docker_server(values)
   check_dir_exists(values, values)
-  values['tftpdir'] = values['exportdir']+"/tftpboot"
-  information_message(values, "Checking TFTP directory")
+  values['tftpdir'] = "#{values['exportdir']}/tftpboot"
+  information_message(values, 'Checking TFTP directory')
   check_dir_exists(values, values['tftpdir'])
   check_dir_owner(values, values['tftpdir'], values['uid'])
-  service_dir = values['tftpdir']+"/"+values['service']
+  service_dir = "#{values['tftpdir']}/#{values['service']}"
   check_dir_exists(values, service_dir)
   exists = check_docker_image_exists(values['scriptname'])
   if exists == false
@@ -17,35 +19,31 @@ def configure_docker_server(values)
     check_dir_exists(values, $docker_host_tftp_dir)
     check_dir_exists(values, $docker_host_apache_dir)
     check_dir_exists(values, $docker_host_file_dir)
-    docker_file = $docker_host_file_dir+"/Dockerfile"
+    docker_file = "#{$docker_host_file_dir}/Dockerfile"
     create_docker_file(docker_file)
     build_docker_file(docker_file)
   else
     verbose_message(values, "Docker image #{values['scriptname']} already exists")
   end
-  return
+  nil
 end
 
 # Configure Dockerfile
 
 def create_docker_file(values, docker_file)
-  if File.exist?(docker_file)
-    File.delete(docker_file)
-  end
-  file = File.open(docker_file, "w")
+  File.delete(docker_file) if File.exist?(docker_file)
+  file = File.open(docker_file, 'w')
   file.write("FROM: centos:latest\n")
   file.write("VOLUME #{values['tftpdir']} #{values['apachedir']}")
-  file.write("RUN yum install -y tftp-server syslinux wget apache")
-  file.write("EXPOSE 69/udp")
-  file.write("ENTRYPOINT in.tftpd -s /srv -4 -L -a 0.0.0.0:69")
-  return
+  file.write('RUN yum install -y tftp-server syslinux wget apache')
+  file.write('EXPOSE 69/udp')
+  file.write('ENTRYPOINT in.tftpd -s /srv -4 -L -a 0.0.0.0:69')
+  nil
 end
 
 # Build Dockerfile
 
-def build_docker_file(values, docker_file)
-  message = "Building: #{docker_file}"
-  command = "cd #{$docker_host_base_dir} ; docker build --tag #{values['scriptname']} #{values['scriptname']}"
-  #execute_command(values, message, command)
-  return
+def build_docker_file(_values, _docker_file)
+  # execute_command(values, message, command)
+  nil
 end
